requirements.txt
google-adk>=1.0.0
sqlalchemy>=2.0.0
pg8000>=1.30.0
gspread>=6.0.0
google-auth>=2.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
